requirements.txt
pandas>=1.5.3
psycopg2-binary>=2.9.6
matplotlib>=3.6.0
seaborn>=0.12.0
scipy>=1.10.0
statsmodels>=0.13.5